feat(modifiers): Add drunken modifiers (unfinished)

+++ api/src/main/java/dev/jsinco/brewery/api/util/WeightedProbabilityElement.java
@@ -2,8 +2,5 @@ package dev.jsinco.brewery.api.util;
 
 public interface WeightedProbabilityElement {
 
-    /**
-     * @return A probability weight
-     */
-    int probabilityWeight();
+    double probabilityWeight();
 }
